// src/client/src/models/Need.ts
// @ts-ignore
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * People in Need - Springboard Service
 * For the backend supporting the Citi 2023 Techfest Hackathon project.  This backend is meant to be a simple CRUD service for supporting  collaboration between organizations and the beneficiaries supported by them.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists } from '../runtime';
/**
 * 
 * @export
 * @interface Need
 */
export interface Need {
    /**
     * 
     * @type {number}
     * @memberof Need
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Need
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Need
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof Need
     */
    alternativeName?: string;
}

/**
 * Check if a given object implements the Need interface.
 */
export function instanceOfNeed(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function NeedFromJSON(json: any): Need {
    return NeedFromJSONTyped(json, false);
}

export function NeedFromJSONTyped(json: any, ignoreDiscriminator: boolean): Need {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'alternativeName': !exists(json, 'alternativeName') ? undefined : json['alternativeName'],
    };
}

export function NeedToJSON(value?: Need | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'description': value.description,
        'alternativeName': value.alternativeName,
    };
}